Added the testing files and my kNN class
I have basically all the framework done for my class but for some reason
the accuracy is at 50% for all of them. Gonna look at it in the morning
with fresh eyes. Also commented out the preprocess stuff cause we don't
need to run it as were building and testing our algorithms. Just a waste
of time right now